requirements.txt
Sphinx >= 1.7.0
Unidecode >= 0.4.21
sphinx-autodoc-typehints >= 1.3.0
sphinx-rtd-theme >= 0.2.4
black
isort
pytest-cov
